add script to jumpstart iOS debugging
In devkit v1, basil would send a datagram to the NativeInspector server
containing the iOS device IP address. NativeInspector then connects and
attempts to reconnect to the iOS device until it is turned off. The
script added here, `connect-ios.js`, will alert the NativeInspector
server with an IP address provided on the command line.